demo for BinlogEvent
